Fixed renaming typo file name
