+++ src/App.js
@@ -1,5 +1,5 @@
 import AccountCard from "./shared/components/AccountCard/AccountCard";
-import EditProfileCard from "./shared/components/EditProfieCard/EditProfieCard";
+import EditProfileCard from "./shared/components/EditProfileCard/EditProfileCard";
 
 function App() {
   return (
